% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellTypeDEPlot.R
\name{cellTypeDEPlot}
\alias{cellTypeDEPlot}
\title{Plots DE Heatmap showing how consensus cluster labels are refined by re-clustering over pair-wise DE genes}
\usage{
cellTypeDEPlot(
  dataMatrix,
  nodg = NULL,
  cellTree,
  clusterLabels,
  dynamicColorsList,
  geneLabels = NULL,
  colScheme = "green",
  filename = "DE_Heatmap"
)
}
\arguments{
\item{dataMatrix}{the log-transformed and normalized scRNAseq genes x cells matrix}

\item{nodg}{vector of NODG to visualize cell sequencing differences}

\item{cellTree}{tree obtained from hierarchical clustering}

\item{clusterLabels}{consensus cluster labels}

\item{dynamicColorsList}{labels of newly defined clusters on DE genes}

\item{geneLabels}{labels for the DE genes, here NULL.}

\item{colScheme}{c("green", "blue", "lightblue")}

\item{filename}{name of DE heatmap file}
}
\description{
Plots DE Heatmap showing how consensus cluster labels are refined by re-clustering over pair-wise DE genes
}
\author{
ranjanb
}
