% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scConsensus.R
\name{plotContingencyTable}
\alias{plotContingencyTable}
\title{plots the contingency table of two clustering results}
\usage{
plotContingencyTable(cluster_labels_1 = NULL, cluster_labels_2 = NULL,
  filename = "Contingency_Table.pdf")
}
\arguments{
\item{cluster_labels_1}{First vector of cluster labels for each cell}

\item{cluster_labels_2}{Second vector of cluster labels for each cell}

\item{filename}{name of contingency table file}
}
\description{
plots the contingency table of two clustering results
}
\author{
ranjanb
}
