% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reclusterDEConsensusFast.R
\name{reclusterDEConsensusFast}
\alias{reclusterDEConsensusFast}
\title{Recluster consensus clusters using improved DE gene analysis}
\usage{
reclusterDEConsensusFast(
  dataMatrix,
  consensusClusterLabels,
  method = "wilcox",
  qValThrs = 0.1,
  logFCThrs = 0.5,
  deepSplitValues = 1:4,
  minClusterSize = 10,
  minPerCent = 20,
  filename = "de_gene_object.rds",
  plotName = "DE_Heatmap",
  NumbertopDEGenes = 30,
  nCores = 1
)
}
\arguments{
\item{dataMatrix}{the log-transformed and normalized scRNAseq genes x cells matrix}

\item{consensusClusterLabels}{consensus cell type labels for each cell}

\item{method}{Method used for DE gene statistical test}

\item{qValThrs}{maximum q-value threshold for statistical test}

\item{deepSplitValues}{vector of WGCNA tree cutting deepsplit parameters}

\item{minClusterSize}{specifies the type of minimum cluster size}

\item{minPerCent}{Minium percentage of cells expressing a gene within a cluster}

\item{filename}{name of DE gene object file}

\item{plotName}{name of DE Heatmap file}

\item{NumbertopDEGenes}{number of top DE genes to consider for final clustering}

\item{nCores}{nuber of cores to use for DE gene calculation}

\item{fcThrs}{minimum fold-change threshold for DE gene criterion (natural log as by Seurat convention)}
}
\value{
list containing vector of DE genes, clustering tree and dynamic color list.
}
\description{
Recluster consensus clusters using improved DE gene analysis
}
\author{
ranjanb, schmidtf
}
